+++ notes/admin.py
@@ -6,7 +6,6 @@ class NotebookAdmin(admin.ModelAdmin):
     list_display = ('title', 'slug', 'author', 'created', 'modified')
     search_fields = ('title', 'author__username')
     prepopulated_fields = {'slug': ('title',)}
-    list_filter = ('created', 'modified', 'author')
     ordering = ('-created',)
 
 @admin.register(Note)
@@ -14,5 +13,4 @@ class NoteAdmin(admin.ModelAdmin):
     list_display = ('title', 'slug', 'author', 'created', 'modified', 'notebook')
     search_fields = ('title', 'author__username', 'content')
     prepopulated_fields = {'slug': ('title',)}
-    list_filter = ('created', 'modified', 'author', 'notebook')
     ordering = ('-created',)
