+++ notes/templates/base.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>{% block title %}My Django Site{% endblock %}</title>
+        <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css" rel="stylesheet">
+    </head>
+<body class="bg-gray-100">
+
+    <nav class="bg-blue-400 border-b border-blue-400 p-2">
+        <div class="container mx-auto flex items-center justify-between">
+            <a href="{% url 'note:home' %}" class="text-white text-xl font-bold hover:text-blue-200 transition duration-300">Home</a>
+            
+            <div class="flex items-center space-x-4">
+                <span class="text-white font-semibold text-lg">USER</span>
+                <div class="relative">
+                    
+                    <button id="user-menu-button" class="text-white font-semibold text-lg px-4 py-2 bg-blue-500 hover:bg-blue-700 rounded-md transition duration-300">
+                        User Menu
+                        <svg class="inline w-4 h-4 ml-2" fill="currentColor" viewBox="0 0 20 20">
+                            <path d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 011.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"/>
+                        </svg>
+                    </button>
+
+                    <div id="user-menu" class="hidden absolute right-0 mt-2 w-48 bg-white border border-gray-200 rounded-lg shadow-lg">
+                        <a href="#" class="block px-4 py-2 text-gray-800 hover:bg-blue-100">My Account</a>
+                        <a href="#" class="block px-4 py-2 text-gray-800 hover:bg-blue-100">Change Password</a>
+                        <a href="#" class="block px-4 py-2 text-gray-800 hover:bg-blue-100">Logout</a>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </nav>
+
+    <main class="container mx-auto p-4">
+        {% block content %}{% endblock %}
+    </main>
+
+    <footer class="bg-blue-400 text-white py-4 mt-8">
+        <div class="container mx-auto text-center">
+            <p class="text-sm">&copy; 2024 TestPress. All rights reserved.</p>
+        </div>
+    </footer>
+
+    <script>
+        document.getElementById("user-menu-button").addEventListener("click", function() {
+            const menu = document.getElementById("user-menu");
+            menu.classList.toggle("hidden");
+        });
+    </script>
+</body>
+</html>
