+++ notes/templates/notes/note_detail.html
@@ -1,0 +1,66 @@
+{% extends 'base.html' %}
+
+{% block title %}Note Details{% endblock %}
+
+{% block content %}
+    <div class="max-w-4xl mx-auto bg-white p-6 rounded-lg shadow-lg mt-6 transition-transform transform  hover:shadow-2xl" style="transition-duration: .35s;">
+        <h2 class="text-3xl font-semibold mb-4 text-blue-600">{{ note.title }}</h2>
+        <p class="text-lg mb-2"><strong>Author:</strong> <span class="text-gray-700">{{ note.author.username }}</span></p>
+        <p class="text-lg mb-2"><strong>Notebook:</strong> <span class="text-gray-700">{{ note.notebook.title }}</span></p>
+        <p class="text-lg mb-2"><strong>Slug:</strong> <span class="text-gray-700">{{ note.slug }}</span></p>
+        <p class="text-lg mb-2"><strong>Created:</strong> <span class="text-gray-700">{{ note.created }}</span></p>
+        <p class="text-lg mb-4"><strong>Last Modified:</strong> <span class="text-gray-700">{{ note.modified }}</span></p>
+        <hr class="my-6 border-gray-300">
+
+        <h3 class="text-2xl font-semibold mb-4 text-gray-800">Content</h3>
+        <p class="text-lg text-gray-700">{{ note.content }}</p>
+        <h3 class="text-2xl font-semibold mt-6 mb-4 text-gray-800">Tags</h3>
+        
+        {% if note.tags.all %}
+            <p class="text-gray-700">
+                <strong>Tags:</strong> 
+                {% for tag in note.tags.all %}
+                    <span class="bg-blue-100 text-blue-800 px-2 py-1 rounded-full text-sm">
+                        <a href="{% url "note:similar_note" note.id tag.slug  %}">{{ tag.name }}</a>
+                    </span>&nbsp;
+                {% endfor %}
+            </p>
+        {% else %}
+            <p class="text-gray-600">No tags available for this note.</p>
+        {% endif %}
+
+        <div class="mt-4 border-t pt-4 flex justify-between">
+
+            <a href="{% url 'note:notebook_detail' note.notebook.pk %}">
+                <button class="bg-blue-500 text-white px-6 py-2 rounded-lg shadow-md hover:bg-blue-600 transition duration-300 transform hover:scale-105">
+                    Back to Notebook
+                </button>
+            </a> 
+
+            <div class="flex space-x-4">
+                <a href="#">
+                    <button class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600 transition">
+                        Edit
+                    </button>
+                </a>
+                <form action="#" method="post" class="inline-block">
+                    {% csrf_token %}
+                    <button type="submit" class="bg-red-500 text-white px-4 py-2 rounded hover:bg-red-600 transition">
+                        Delete
+                    </button>
+                </form>
+            </div>
+        </div>
+    </div>
+
+    {% if tag_slug %}
+        {% if similar_note %}
+            {% include "notes/similar_notes.html" %}
+        {% else %}
+            <p class="text-gray-600 mt-8">No similar notes available for this tag.</p>
+    {% endif %}
+{% endif %}
+
+
+
+{% endblock %}
