+++ notes/templates/notes/note_list.html
@@ -1,0 +1,60 @@
+{% extends 'base.html' %}
+
+{% block title %}Notes for {{ notebook.title }}{% endblock %}
+
+{% block content %}
+<h2 class="text-2xl font-semibold mb-6 flex justify-between items-center">
+    <span>Notes of {{ notebook.title }}</span>
+    <a href="#" class="bg-green-500 text-white px-3 text-base py-3 rounded hover:bg-green-600 transition">
+        + New Note
+    </a>
+</h2>
+
+{% if notes %}
+    <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+        {% for note in notes %}
+            <div class="border rounded-lg shadow-lg p-6 bg-white transition-transform transform hover:scale-105 hover:shadow-2xl" style="transition-duration: .35s;">
+                <h3 class="text-xl font-bold mb-2">
+                    <a href="{% url 'note:note_detail' note.pk %}" class="text-blue-600 hover:text-blue-800 transition">{{ note.title }}</a>
+                </h3>
+                <p class="text-gray-600 mb-4"><strong>Content:</strong> {{ note.content|truncatewords:20 }}</p>
+                <p class="text-gray-500 mb-2"><strong>Created:</strong> {{ note.created }}</p>
+                <p class="text-gray-500 mb-2"><strong>Last Modified:</strong> {{ note.modified }}</p>
+                
+                {% if note.tags.all %}
+                    <p class="text-gray-700 mb-2">
+                        <strong>Tags:</strong> 
+                        {% for tag in note.tags.all %}
+                            <span class="bg-blue-100 text-blue-800 px-2 py-1 rounded-full text-sm">{{ tag }}</span>{% if not forloop.last %} {% endif %}
+                        {% endfor %}
+                    </p>
+                {% endif %}
+
+                <div class="mt-4 border-t pt-4 flex justify-between">
+                    <a href="{% url 'note:note_detail' note.pk %}">
+                        <button class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600 transition">
+                            View
+                        </button>
+                    </a> 
+                    <div class="flex space-x-4">
+                        <a href="#">
+                            <button class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600 transition">
+                                Edit
+                            </button>
+                        </a>
+                        <form action="#" method="post" class="inline-block">
+                            {% csrf_token %}
+                            <button type="submit" class="bg-red-500 text-white px-4 py-2 rounded hover:bg-red-600 transition">
+                                Delete
+                            </button>
+                        </form>
+                    </div>
+                </div>
+            </div>
+        {% endfor %}
+    </div>
+{% else %}
+    <p class="text-gray-600">No notes available for this notebook.</p>
+{% endif %}
+
+{% endblock %}
