+++ notes/templates/notes/notebook_list.html
@@ -1,0 +1,55 @@
+{% extends 'base.html' %}
+
+{% block title %}Notebooks{% endblock %}
+
+{% block content %}
+
+<h2 class="text-2xl font-semibold mb-6 flex justify-between items-center">
+    <span>Notebooks</span>
+    <a href="#" class="bg-green-500 text-white px-3 text-base py-3 rounded hover:bg-green-600 transition">
+        +New NoteBook
+    </a>
+</h2>
+
+{% if notebook %}
+    <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+        {% for item in notebook %}
+            <div class="border rounded-lg shadow-lg p-6 transition-transform transform hover:scale-105 hover:shadow-2xl" style="transition-duration: .35s;">
+                <h3 class="text-xl font-bold mb-2 text-blue-600">{{ item.title }}</h3>
+                <p class="text-gray-700 mb-1"><strong>Author:</strong> {{ item.author.username }}</p>
+                <p class="text-gray-700 mb-1"><strong>No. of Notes:</strong> {{ item.notes.count }}</p>
+                <p class="text-gray-500 mb-1"><strong>Created:</strong> {{ item.created }}</p>
+                <p class="text-gray-500"><strong>Last Modified:</strong> {{ item.modified }}</p>
+
+                <div class="mt-4 border-t pt-4 flex justify-between">
+                    <a href="{% url 'note:notebook_detail' item.pk %}">
+                        <button class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600 transition">
+                            View Notes
+                        </button>
+                    </a> 
+                    <div class="flex space-x-4">
+                        <form action="#" method="post">
+                            {% csrf_token %}
+                            <button type="submit" class="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600 transition">
+                                Edit
+                            </button>
+                        </form>
+                
+                        <form action="#" method="post" class="inline-block">
+                            {% csrf_token %}
+                            <button type="submit" class="bg-red-500 text-white px-4 py-2 rounded hover:bg-red-600 transition">
+                                Delete
+                            </button>
+                        </form>
+                
+                    </div>
+                </div>
+                
+            </div>
+        {% endfor %}
+    </div>
+{% else %}
+    <p class="text-gray-600">No notebooks available.</p>
+{% endif %}
+
+{% endblock %}
