+++ notes/templates/notes/similar_notes.html
@@ -1,0 +1,30 @@
+<div class="mt-8 bg-gray-50 p-6 rounded-lg shadow-lg">
+    <h3 class="text-2xl font-semibold text-gray-800 mb-6">Similar Notes</h3>
+
+    <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+        {% for note in similar_note %}
+            <div class="bg-white p-6 rounded-lg shadow-md hover:shadow-xl transition-shadow">
+                <h4 class="text-xl font-semibold text-blue-600 mb-4">
+                    <a href={% url 'note:note_detail' note.id%}>{{ note.title }}</a>
+                </h4>
+                <p class="text-lg mb-2 text-gray-700"><strong>Author:</strong> {{ note.author.username }}</p>
+                <p class="text-lg mb-2 text-gray-700"><strong>Notebook:</strong> {{ note.notebook.title }}</p>
+                <p class="text-lg mb-2 text-gray-700"><strong>Slug:</strong> {{ note.slug }}</p>
+                <p class="text-lg mb-4 text-gray-500"><strong>Created:</strong> {{ note.created }}</p>
+                <p class="text-lg mb-4 text-gray-600">
+                    <strong>Content:</strong> 
+                    <span class="block text-sm text-gray-700">{{ note.content|truncatewords:20 }}</span>
+                </p>
+
+                <div class="text-lg text-gray-700">
+                    <strong>Tags:</strong>
+                    {% for tag in note.tags.all %}
+                        <span class="bg-blue-100 text-blue-800 px-2 py-1 rounded-full text-sm mr-2 mt-2 inline-block">
+                            {{ tag.name }}
+                        </span>
+                    {% endfor %}
+                </div>
+            </div>
+        {% endfor %}
+    </div>
+</div>
