+++ notes_app/urls.py
@@ -18,6 +18,6 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
-    # path('', include('notes.urls')),
+    path('', include('notes.urls')),
     path('admin/', admin.site.urls),
 ]
